+++ src/main/java/Grammar/Plan/BinaryPlan.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 public record BinaryPlan(Plan p1, Plan p2) implements Plan {
     @Override
-    public boolean eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError {
+    public boolean eval(Map<String, Long> bindings, Player p, Territory t) throws EvalError {
         if (p1.eval(bindings, p ,t)) {
             return p2.eval(bindings, p, t);
         }
